+++ modules/terraform-aci-remote-leaf-pod-redundancy-policy/main.tf
@@ -3,7 +3,7 @@ resource "aci_rest_managed" "infraRlPodRedPol" {
   class_name = "infraRlPodRedPol"
   content = {
     name                   = "default"
-    enableRlPodRedPol      = var.enable_remote_leaf_policy == true ? "yes" : "no"
+    enableRlPodRedPol      = var.enable_policy == true ? "yes" : "no"
     enablePodRedPreemption = var.preemption == true ? "yes" : "no"
   }
 }